pyproject.toml
[project]
name = "production-blueprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
